Initial version of structure removal.  Passes 1 test!
Currently we don't support iterators, as supporting them requires more
type information than we have in some case.  This should be fixable:
we can support them at least when we have the type information,
and also we should be able to reconstruct the missing type information
but we don't do type checking at the moment.