Don't check input modes, or at least not the way we were doing.
Previously, we insisted that assertions only mention inputs, the
reasoning being that this is what callers have control over.
However, some examples also use locals in assertions, which makes sense
as long as the local's definitions only depend on inputs.

For the time being, we simply don't do this check, and allow assertions
to depend on whatever they need to.