Remove `onlyInputs` and add comments.
The idea with `onlyInputs` was that assumptions should only mention things
in the inputs.  However, sometimes people define local variables to
compute some relation between inputs, and assumptions can only use those.

At the end of the day, everything in a node only depends on the inputs,
so we just remove this check for now.

XXX: One possible exception is previous values of variables? Not sure.